Display binary content in Example Connector
This PR modifies the Example Connector to show base64 encoded binary content under the _attachment field. This field can be hardcoded, point to a file path or point to a list of directories (as shown in this PR). If a list of directories is provided, each individual path yields a separate document