Feature: Add simple rule validation support and implement basic generic validation (#467)
Add simple rule validation support to the framework.

Connector implementations can plug-in dedicated validation classes for different stages (pre-processing, post-processing or for all stages). The framework will take care of the rest (instantiating the correct classes, passing the advanced snippet or the simple rules and checking, if pre-processing simple rules are active).

# Conflicts:
#	lib/connectors/mongodb/mongo_rules_parser.rb
#	lib/core/filtering/hash_against_schema_validator.rb
#	spec/connectors/mongodb/mongo_advanced_snippet_against_schema_validator_spec.rb
#	spec/core/filtering/hash_against_schema_validator_spec.rb