Fix: Implementation for MongoDB simple rules pre-processing for 'starts_with', 'ends_with' and 'contains' missing (#458) (#486)
Add the missing implementation for MongoDB pre-processing simple rules for the rules 'starts_with', 'ends_with' and 'contains'. This fixes the bug that a user can select these rules in the UI, but a corresponding sync won't be started as the backend throws an error that the selected rule is unknown.